+++ src/client/java/prismatic/shards/stellarity/datagen/AdvancementProvider.java
@@ -1,40 +1,45 @@
 package prismatic.shards.stellarity.datagen;
 
+import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricAdvancementProvider;
-import net.minecraft.advancements.Advancement;
-import net.minecraft.advancements.AdvancementHolder;
-import net.minecraft.advancements.AdvancementRequirements;
-import net.minecraft.advancements.AdvancementType;
+import net.minecraft.advancements.*;
 import net.minecraft.advancements.predicates.*;
 import net.minecraft.advancements.predicates.entity.EntityPredicate;
 import net.minecraft.advancements.predicates.entity.EntityTypePredicate;
 import net.minecraft.advancements.predicates.entity.PlayerPredicate;
 import net.minecraft.advancements.triggers.*;
+import net.minecraft.commands.arguments.NbtPathArgument;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.HolderSet;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.Identifier;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.entity.EntityTypes;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.predicates.LootItemBlockStatePropertyCondition;
 import net.minecraft.world.level.storage.loot.predicates.LootItemEntityPropertyCondition;
 import org.jspecify.annotations.NonNull;
 import prismatic.shards.stellarity.Stellarity;
+import prismatic.shards.stellarity.datagen.tags.BiomeTagProvider;
 import prismatic.shards.stellarity.registry.StellarityBlocks;
 import prismatic.shards.stellarity.registry.StellarityCriteriaTriggers;
 import prismatic.shards.stellarity.registry.StellarityItems;
 import prismatic.shards.stellarity.registry.criterion_trigger.DashTrigger;
 import prismatic.shards.stellarity.registry.criterion_trigger.SpecialCraftTrigger;
 import prismatic.shards.stellarity.registry.criterion_trigger.VoidFishedTrigger;
+import prismatic.shards.stellarity.registry.entity_sub_predicate.EntityAttributeModifiersPredicate;
+import prismatic.shards.stellarity.registry.entity_sub_predicate.NbtNumberPredicate;
 import prismatic.shards.stellarity.tags.StellarityDamageTypeTags;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
+import java.util.stream.Stream;
 
 
 public class AdvancementProvider extends FabricAdvancementProvider {
@@ -71,18 +76,19 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 
 	@Override
 	public void generateAdvancement(HolderLookup.Provider registryLookup, @NonNull Consumer<AdvancementHolder> consumer) {
-		var items = registryLookup.lookupOrThrow(Registries.ITEM);
-		var entities = registryLookup.lookupOrThrow(Registries.ENTITY_TYPE);
+		final var items = registryLookup.lookupOrThrow(Registries.ITEM);
+		final var entities = registryLookup.lookupOrThrow(Registries.ENTITY_TYPE);
 
 
-		var ENTER_END_GATEWAY = dummy(Stellarity.mcId("end/enter_end_gateway"));
-		var ENTER_END = dummy(Stellarity.mcId("story/enter_the_end"));
-		var END_ROOT = dummy(Stellarity.mcId("end/root"));
-		var KILL_DRAGON = dummy(Stellarity.mcId("end/kill_dragon"));
-		var ADVENTURE_ROOT = dummy(Stellarity.mcId("adventure/root"));
+		final var ENTER_END_GATEWAY = dummy(Stellarity.mcId("end/enter_end_gateway"));
+		final var ENTER_END = dummy(Stellarity.mcId("story/enter_the_end"));
+		final var END_ROOT = dummy(Stellarity.mcId("end/root"));
+		final var KILL_DRAGON = dummy(Stellarity.mcId("end/kill_dragon"));
+		final var ADVENTURE_ROOT = dummy(Stellarity.mcId("adventure/root"));
+		final var KILL_A_MOB = dummy(Stellarity.mcId("adventure/kill_a_mob"));
 
 
-		var VOID_REELS = advancement()
+		final var VOID_REELS = advancement()
 			.display(StellarityItems.FISHER_OF_VOIDS,
 				Component.translatable("advancements.stellarity.void_reels"), Component.translatable("advancements.stellarity.void_reels.description"),
 				null, TASK, true, true, false
@@ -93,19 +99,27 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			)).requirements(requires(new String[][]{{"fishing"}}))
 			.build(Stellarity.id("void_fishing/void_reels"));
 
-		var TOPPED_OFF = advancement()
+		final var TOPPED_OFF = advancement()
 			.display(
 				StellarityItems.CRYSTAL_HEARTFISH,
 				Component.translatable("advancements.stellarity.topped_off"), Component.translatable("advancements.stellarity.topped_off.description"),
 				null, TASK, true, true, false
 			)
 			.parent(VOID_REELS)
-			// minecraft doesn't bundle easy way, and the only attribute modifier checks the ITEM not the player
-			.addCriterion("impossible", impossible())
+			.addCriterion("eat", CriteriaTriggers.CONSUME_ITEM.createCriterion(new ConsumeItemTrigger.TriggerInstance(
+				Optional.of(EntityPredicate.wrap(EntityPredicate.Builder.entity().put(EntityAttributeModifiersPredicate.CODEC, new EntityAttributeModifiersPredicate(List.of(
+					new EntityAttributeModifiersPredicate.EntryPredicate(
+						HolderSet.direct(Attributes.MAX_HEALTH), Optional.of(Stellarity.id("crystal_heartfish")),
+						MinMaxBounds.Doubles.atLeast(10), Optional.empty()
+					)
+				))))),
+				Optional.of(ItemPredicate.Builder.item().of(items, StellarityItems.CRYSTAL_HEARTFISH).build())
+			)))
+			.requirements(requires("eat"))
 			.build(Stellarity.id("void_fishing/topped_off"));
 
 
-		var FIND_DUSKBERRY = advancement()
+		final var FIND_DUSKBERRY = advancement()
 			.display(
 				StellarityItems.DUSKBERRY,
 				Component.translatable("advancements.stellarity.duskberry_find"), Component.translatable("advancements.stellarity.duskberry_find.description"),
@@ -116,7 +130,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			.requirements(requires("get_item"))
 			.build(Stellarity.id("exploration/duskberry/find"));
 
-		var POOR_LIFE_CHOICES = advancement()
+		final var POOR_LIFE_CHOICES = advancement()
 			.display(
 				StellarityItems.DUSKBERRY,
 				Component.translatable("advancements.stellarity.poor_life_choices"), Component.translatable("advancements.stellarity.poor_life_choices.description"),
@@ -131,7 +145,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			.requirements(requires("eat", "feed", "place"))
 			.build(Stellarity.id("exploration/duskberry/poor_life_choices"));
 
-		var SACRIFICAL_RITUAL = advancement().display(
+		final var SACRIFICAL_RITUAL = advancement().display(
 				Items.END_CRYSTAL,
 				Component.translatable("advancements.stellarity.sacrificial_ritual"), Component.translatable("advancements.stellarity.sacrificial_ritual.description"),
 				null, GOAL, true, true, false
@@ -140,7 +154,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			.requirements(requires("summon"))
 			.build(Stellarity.id("ender_dragon/sacrificial_ritual"));
 
-		var RESPAWN_DRAGON = advancement().display(
+		final var RESPAWN_DRAGON = advancement().display(
 				Items.END_CRYSTAL,
 				Component.translatable("advancements.end.respawn_dragon.title"), Component.translatable("advancements.end.respawn_dragon.description"),
 				null, TASK, true, true, false
@@ -153,9 +167,9 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			)))
 			.requirements(
 				requires("summon")
-			).build(Stellarity.mcId("end/respawn_dragon"));
+			).build(Stellarity.id("end/respawn_dragon"));
 
-		var ALTAR_OF_THE_ACCURSED_INTRO = advancement().display(
+		final var ALTAR_OF_THE_ACCURSED_INTRO = advancement().display(
 				StellarityItems.ENDONOMICON,
 				Component.translatable("advancements.stellarity.altar_of_the_accursed_intro"), Component.translatable("advancements.stellarity.altar_of_the_accursed_intro.description"),
 				null, GOAL, true, true, false
@@ -166,10 +180,10 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 				Optional.of(ItemPredicate.Builder.item().of(items, StellarityItems.ENDONOMICON).build())
 			))
 			.requirements(requires("craft"))
-			.build(Stellarity.mcId("altar_of_the_accursed/altar_of_the_accursed_intro"));
+			.build(Stellarity.id("altar_of_the_accursed/altar_of_the_accursed_intro"));
 
 		// TODO: reparent to intro to dark magic
-		var CURSED_CRAFTING = advancement().display(
+		final var CURSED_CRAFTING = advancement().display(
 				StellarityItems.ALTAR_OF_THE_ACCURSED,
 				Component.translatable("advancements.stellarity.cursed_crafting"), Component.translatable("advancements.stellarity.cursed_crafting.description"),
 				null, GOAL, true, true, false
@@ -182,9 +196,9 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 				Optional.empty()
 			))
 			.requirements(requires("craft"))
-			.build(Stellarity.mcId("altar_of_the_accursed/cursed_crafting"));
+			.build(Stellarity.id("altar_of_the_accursed/cursed_crafting"));
 
-		var CRAFT_FULL_SHULKER_ARMOR = advancement().display(
+		final var CRAFT_FULL_SHULKER_ARMOR = advancement().display(
 				StellarityItems.SHULKER_CHESTPLATE,
 				Component.translatable("advancements.stellarity.craft_full_shulker_armor"), Component.translatable("advancements.stellarity.craft_full_shulker_armor.description"),
 				null, CHALLENGE, true, true, false
@@ -194,9 +208,9 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			.addCriterion("leggings", InventoryChangeTrigger.TriggerInstance.hasItems(StellarityItems.SHULKER_LEGGINGS))
 			.addCriterion("boots", InventoryChangeTrigger.TriggerInstance.hasItems(StellarityItems.SHULKER_BOOTS))
 			.requirements(requires("helmet", "chestplate", "leggings", "boots"))
-			.build(Stellarity.mcId("altar_of_the_accursed/craft_full_shulker_armor"));
+			.build(Stellarity.id("altar_of_the_accursed/craft_full_shulker_armor"));
 
-		var ELECTRIFIED = advancement().display(
+		final var ELECTRIFIED = advancement().display(
 				StellarityItems.COPPER_ELEKTRA_SHIELD,
 				Component.translatable("advancements.stellarity.electrified"), Component.translatable("advancements.stellarity.electrified.description"),
 				null, TASK, true, true, false
@@ -208,7 +222,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			.requirements(requires("dash"))
 			.build(Stellarity.id("adventure/electrified"));
 
-		var BLOOD_FOR_BLOOD = advancement().display(
+		final var BLOOD_FOR_BLOOD = advancement().display(
 				StellarityItems.TAMARIS,
 				Component.translatable("advancements.stellarity.blood_for_blood"), Component.translatable("advancements.stellarity.blood_for_blood.description"),
 				null, TASK, true, true, false
@@ -219,9 +233,50 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 			.requirements(requires("kill"))
 			.build(Stellarity.id("altar_of_the_accursed/blood_for_blood"));
 
+		final var NIGHT_SKY_STALKERS = advancement().display(
+				Items.PHANTOM_MEMBRANE,
+				Component.translatable("advancements.stellarity.night_sky_stalkers"), Component.translatable("advancements.stellarity.night_sky_stalkers.description"),
+				null, TASK, true, true, false
+			).parent(KILL_A_MOB)
+			.addCriterion("kill", KilledTrigger.TriggerInstance.playerKilledEntity(new EntityPredicate.Builder().of(entities, EntityTypes.PHANTOM)))
+			.requirements(requires("kill"))
+			.build(Stellarity.id("adventure/night_sky_stalkers"));
+
+		Map<NbtPathArgument.NbtPath, MinMaxBounds.Doubles> largePhantomData = new HashMap<>();
+		try {
+			largePhantomData.put(NbtPathArgument.NbtPath.of("size"), MinMaxBounds.Doubles.atLeast(2));
+		} catch (CommandSyntaxException e) {
+			throw new RuntimeException(e);
+		}
+		final var KILL_LARGE_PHANTOM = advancement().display(
+				Items.PHANTOM_SPAWN_EGG,
+				Component.translatable("advancements.stellarity.kill_large_phantom"), Component.translatable("advancements.stellarity.kill_large_phantom.description"),
+				null, TASK, true, true, false
+			).parent(NIGHT_SKY_STALKERS)
+			.addCriterion("kill_large", KilledTrigger.TriggerInstance.playerKilledEntity(
+				new EntityPredicate.Builder().of(entities, EntityTypes.PHANTOM).put(NbtNumberPredicate.CODEC, new NbtNumberPredicate(largePhantomData))
+			))
+			.requirements(requires("kill_large"))
+			.build(Stellarity.id("adventure/kill_large_phantom"));
+
+		var DISCOVER_ALL_BIOMES_BUILDER = advancement().display(
+			Items.LEATHER_BOOTS,
+			Component.translatable("advancements.stellarity.discover_all_biomes"), Component.translatable("advancements.stellarity.discover_all_biomes.description"),
+			null, CHALLENGE, true, true, false
+		).parent(ENTER_END_GATEWAY).rewards(new AdvancementRewards.Builder().addExperience(4500).build());
+		List<String> biomes = new ArrayList<>();
+		for (var biome : (Iterable<ResourceKey<Biome>>) () -> Stream.concat(BiomeTagProvider.outerVanilla(), BiomeTagProvider.stellarity()).iterator()) {
+			var name = biome.identifier().getPath();
+			biomes.add(name);
+			DISCOVER_ALL_BIOMES_BUILDER.addCriterion(name, PlayerTrigger.TriggerInstance.located(new LocationPredicate.Builder().setBiomes(HolderSet.direct(registryLookup.getOrThrow(biome)))));
+		}
+		DISCOVER_ALL_BIOMES_BUILDER.requirements(requires(biomes.toArray(String[]::new)));
+
+
 		for (var advancement : List.of(
 			VOID_REELS, TOPPED_OFF, FIND_DUSKBERRY, POOR_LIFE_CHOICES, SACRIFICAL_RITUAL, RESPAWN_DRAGON,
-			CURSED_CRAFTING, CRAFT_FULL_SHULKER_ARMOR, ALTAR_OF_THE_ACCURSED_INTRO, ELECTRIFIED, BLOOD_FOR_BLOOD
+			CURSED_CRAFTING, CRAFT_FULL_SHULKER_ARMOR, ALTAR_OF_THE_ACCURSED_INTRO, ELECTRIFIED, BLOOD_FOR_BLOOD,
+			NIGHT_SKY_STALKERS, KILL_LARGE_PHANTOM
 		)) {
 			consumer.accept(advancement);
 		}