+++ src/client/java/prismatic/shards/stellarity/datagen/tags/BiomeTagProvider.java
@@ -5,11 +5,15 @@ import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagsProvider;
 import net.fabricmc.fabric.api.tag.convention.v2.ConventionalBiomeTags;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.tags.BiomeTags;
 import net.minecraft.world.level.biome.Biome;
 import org.jspecify.annotations.NonNull;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
+import java.util.stream.Stream;
 
 import static net.minecraft.world.level.biome.Biomes.*;
 import static prismatic.shards.stellarity.key.StellarityBiomes.*;
@@ -21,13 +25,21 @@ public class BiomeTagProvider extends FabricTagsProvider<Biome> {
 		super(output, Registries.BIOME, registriesFuture);
 	}
 
+	public static Stream<ResourceKey<Biome>> stellarity() {
+		return Stream.of(AMETHYST_FOREST, ASHFALL_DELTAS, CRYSTAL_CRAGS, END_SHRUBLAND, END_WILDS, ENDER_WASTES, ENDLESS_DUNES, FIERY_HILLS, FLESH_TUNDRA, FROSTED_VALLEY, FROZEN_MARSH, FROZEN_SHRUBLAND, FROZEN_SPIKES, HALLOWED_TUNDRA, PRISMARINE_FOREST, THE_HALLOW, THE_NEST, WARPED_MARSH);
+	}
+
+	public static Stream<ResourceKey<Biome>> outerVanilla() {
+		return Stream.of(END_BARRENS, END_HIGHLANDS, END_MIDLANDS, SMALL_END_ISLANDS);
+	}
+
 
 	@SuppressWarnings("unchecked")
 	@Override
 	protected void addTags(HolderLookup.@NonNull Provider provider) {
 		builder(SNOWY).add(FROZEN_SPIKES, FROSTED_VALLEY, FROZEN_MARSH, HALLOWED_TUNDRA, FROZEN_SHRUBLAND);
-		builder(ALL_STELLARITY).add(AMETHYST_FOREST, ASHFALL_DELTAS, CRYSTAL_CRAGS, END_SHRUBLAND, END_WILDS, ENDER_WASTES, ENDLESS_DUNES, FIERY_HILLS, FLESH_TUNDRA, FROSTED_VALLEY, FROZEN_MARSH, FROZEN_SHRUBLAND, FROZEN_SPIKES, HALLOWED_TUNDRA, PRISMARINE_FOREST, THE_HALLOW, THE_NEST, WARPED_MARSH);
-		builder(ALL_OUTER).forceAddTag(ALL_STELLARITY).add(END_BARRENS, END_HIGHLANDS, END_MIDLANDS, SMALL_END_ISLANDS);
+		builder(ALL_STELLARITY).addAll(stellarity());
+		builder(ALL_OUTER).forceAddTag(ALL_STELLARITY).addAll(outerVanilla());
 		builder(ConventionalBiomeTags.IS_END).forceAddTag(ALL_STELLARITY);
 		builder(BiomeTags.IS_END).forceAddTag(ALL_STELLARITY);
 		builder(BiomeTags.ALLOWS_TROPICAL_FISH_SPAWNS_AT_ANY_HEIGHT).forceAddTag(ALL_STELLARITY);