+++ src/main/java/prismatic/shards/stellarity/Stellarity.java
@@ -47,6 +47,7 @@ public class Stellarity implements ModInitializer {
 	@Override
 	public void onInitialize() {
 		StellarityNumberProviders.init();
+		StellarityEntitySubPredicates.init();
 		StellarityRegistries.init();
 		StellarityRegistryEntryModifications.init();
 		StellarityFeatures.init();